feat: Create new class to connect with database
create new class `DBClient` that connect to mongodb database. The class contains three methods `isAlive()` to check if the connection to MongoDB is a success or not, `nbUsers()` that returns the number of documents in the collection users, and `nbFiles` that returns the number of documents in the collection files.